+++ frontend/carbook/src/components/Profile/index.ts
@@ -1,0 +1,4 @@
+export { default as HeaderInfo } from "./HeaderInfo";
+export { default as HeaderContents } from "./HeaderContents";
+export { default as Followlists } from "./FollowLists";
+export { default as PostLists } from "./PostLists";
