+++ frontend/carbook/src/pages/Profile/Profile.scss
@@ -54,7 +54,7 @@
           flex-direction: column;
           justify-content: space-around;
 
-          visibility: hidden;
+          // visibility: hidden;
 
           h3 {
             font-weight: 700;
