+++ frontend/carbook/tsconfig.json
@@ -21,7 +21,6 @@
       "@": ["src"],
       "@/*": ["src/*"]
     }
-
   },
   "ts-node": {
     "esm": true
